requirements.txt
openai~=1.40.6
python-dotenv~=1.0.1
datasets~=2.21.0
tiktoken~=0.7.0
